+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## current branch only
+- return OK in state check if no vservers are configured (#44)
+- replaced depreacted calls from Perl-Nagios-Plugin with new names from Perl-Monitoring-Plugin
+- experimental support for objecttypes with a slash inside, e.g. for nspartition/foo (#43)
+- enhanced support for array responses in check_threshold_and_get_perfdata (#43)
+
 ## v1.5.1 (2018-06-10)
 - fixed minor typo, made matches/matches_not support matching in arrays
 - output for array matches more comprehensible (#33)